requirements.txt
requests
defusedxml
pycrypto
